+++ src/neon/neon_radix4.rs
@@ -29,7 +29,7 @@ pub struct NeonRadix4<N: NeonNum, T> {
 impl<N: NeonNum, T: FftNum> NeonRadix4<N, T> {
     /// Constructs a new NeonRadix4 which computes FFTs of size 4^k * base_fft.len()
     #[inline]
-    pub fn new(k: usize, base_fft: Arc<dyn Fft<T>>) -> Self {
+    pub fn new(k: u32, base_fft: Arc<dyn Fft<T>>) -> Self {
         // Internal sanity check: Make sure that S == T.
         // This struct has two generic parameters S and T, but they must always be the same, and are only kept separate to help work around the lack of specialization.
         // It would be cool if we could do this as a static_assert instead
@@ -196,8 +196,8 @@ mod unit_tests {
         }
     }
 
-    fn test_neon_radix4_64_with_base(k: usize, base_fft: Arc<dyn Fft<f64>>) {
-        let len = base_fft.len() * 4usize.pow(k as u32);
+    fn test_neon_radix4_64_with_base(k: u32, base_fft: Arc<dyn Fft<f64>>) {
+        let len = base_fft.len() * 4usize.pow(k);
         let direction = base_fft.fft_direction();
         let fft = NeonRadix4::<f64, f64>::new(k, base_fft);
         check_fft_algorithm::<f64>(&fft, len, direction);
@@ -215,8 +215,8 @@ mod unit_tests {
         }
     }
 
-    fn test_neon_radix4_32_with_base(k: usize, base_fft: Arc<dyn Fft<f32>>) {
-        let len = base_fft.len() * 4usize.pow(k as u32);
+    fn test_neon_radix4_32_with_base(k: u32, base_fft: Arc<dyn Fft<f32>>) {
+        let len = base_fft.len() * 4usize.pow(k);
         let direction = base_fft.fft_direction();
         let fft = NeonRadix4::<f32, f32>::new(k, base_fft);
         check_fft_algorithm::<f32>(&fft, len, direction);
