+++ src/wasm_simd/wasm_simd_planner.rs
@@ -44,7 +44,7 @@ pub enum Recipe {
         inner_fft: Arc<Recipe>,
     },
     Radix4 {
-        k: usize,
+        k: u32,
         base_fft: Arc<Recipe>,
     },
     Butterfly1,
@@ -609,7 +609,7 @@ impl<T: FftNum> FftPlannerWasmSimd<T> {
 
     fn design_radix4(&mut self, len: usize) -> Arc<Recipe> {
         // plan a step of radix4
-        let exponent = len.trailing_zeros() as usize;
+        let exponent = len.trailing_zeros();
         let base_exponent = match exponent {
             0 => 0,
             1 => 1,
